requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
grpcio-reflection==1.76.0
protobuf==6.33.2
xgboost==3.1.2
pandas==2.3.3
numpy==1.24.2